No need for pandaComm folder, installed by setup.py from pandaVis repo
