+++ nab/detectors/htmcore/htmcore_detector.py
@@ -40,6 +40,14 @@ from nab.detectors.base import AnomalyDetector
 # has been seen so far.
 SPATIAL_TOLERANCE = 0.05
 
+PANDA_VIS_ENABLED = False # if we want to run pandaVis tool
+
+if PANDA_VIS_ENABLED:
+    from PandaVis.pandaComm.server import PandaServer
+    from PandaVis.pandaComm.dataExchange import ServerData, dataHTMObject, dataLayer, dataInput
+
+    pandaServer = PandaServer()
+
 parameters_numenta_comparable = {
   # there are 2 (3) encoders: "value" (RDSE) & "time" (DateTime weekend, timeOfDay)
   'enc': {
@@ -114,6 +122,11 @@ class HtmcoreDetector(AnomalyDetector):
     # internal helper variables:
     self.inputs_ = []
     self.iteration_ = 0
+
+    # initialize pandaVis server
+    if PANDA_VIS_ENABLED:
+        pandaServer.Start()
+        self.BuildPandaSystem(parameters_numenta_comparable)
 
 
   def getAdditionalHeaders(self):
@@ -238,8 +251,18 @@ class HtmcoreDetector(AnomalyDetector):
 
       # 3. Temporal Memory
       # Execute Temporal Memory algorithm over active mini-columns.
-      self.tm.compute(activeColumns, learn=True)
+      #self.tm.compute(activeColumns, learn=True)
+
+      self.tm.activateCells(activeColumns, learn=True)
+      # activateDendrites calculates active segments
+      self.tm.activateDendrites(learn=True)
+      # predictive cells are calculated directly from active segments
+      predictiveCells = self.tm.getPredictiveCells()
+
       self.tm_info.addData( self.tm.getActiveCells().flatten() )
+
+      if PANDA_VIS_ENABLED:
+        self.PandaUpdateData(ts, val, valueBits, dateBits , activeColumns, predictiveCells)
 
       # 4.1 (optional) Predictor #TODO optional
       #TODO optional: also return an error metric on predictions (RMSE, R2,...)
@@ -280,4 +303,74 @@ class HtmcoreDetector(AnomalyDetector):
           # print(self.tm_info)
           pass
 
+      if PANDA_VIS_ENABLED:
+          pandaServer.BlockExecution()
+
+
       return (anomalyScore, raw)
+
+  def BuildPandaSystem(self,modelParams):
+      global serverData
+      serverData = ServerData()
+      serverData.HTMObjects["HTM1"] = dataHTMObject()
+      serverData.HTMObjects["HTM1"].inputs["Value"] = dataInput()
+      serverData.HTMObjects["HTM1"].inputs["TimeOfDay"] = dataInput()
+
+      serverData.HTMObjects["HTM1"].layers["Layer1"] = dataLayer(
+          modelParams["sp"]["columnCount"],
+          modelParams["tm"]["cellsPerColumn"],
+      )
+      serverData.HTMObjects["HTM1"].layers["Layer1"].proximalInputs = ["Value","TimeOfDay"]
+      serverData.HTMObjects["HTM1"].layers["Layer1"].distalInputs = ["Layer1"]
+
+
+  def PandaUpdateData(self, timestamp, value, valueSDR, datetimeSDR, activeColumns, predictiveCells):
+
+    pandaServer.currentIteration = self.iteration_ # update server's iteration number
+    # do not update if we are running GOTO iteration command
+    if (not pandaServer.cmdGotoIteration or (
+            pandaServer.cmdGotoIteration and pandaServer.gotoIteration == pandaServer.currentIteration)):
+      # ------------------HTMpandaVis----------------------
+      # fill up values
+      serverData.iterationNo = pandaServer.currentIteration
+      serverData.HTMObjects["HTM1"].inputs["Value"].stringValue = "VALUE:" + str(value)
+      serverData.HTMObjects["HTM1"].inputs["Value"].bits = valueSDR.sparse
+      serverData.HTMObjects["HTM1"].inputs["Value"].count = valueSDR.size
+
+      serverData.HTMObjects["HTM1"].inputs["TimeOfDay"].stringValue = "TIME OF DAY:" + str(timestamp)
+      serverData.HTMObjects["HTM1"].inputs["TimeOfDay"].bits = datetimeSDR.sparse
+      serverData.HTMObjects["HTM1"].inputs["TimeOfDay"].count = datetimeSDR.size
+
+      serverData.HTMObjects["HTM1"].layers["Layer1"].activeColumns = activeColumns.sparse
+
+      serverData.HTMObjects["HTM1"].layers["Layer1"].winnerCells = self.tm.getWinnerCells().sparse
+      serverData.HTMObjects["HTM1"].layers["Layer1"].activeCells = self.tm.getActiveCells().sparse
+      serverData.HTMObjects["HTM1"].layers["Layer1"].predictiveCells = predictiveCells.sparse
+
+      # print("ACTIVECOLS:"+str(serverData.HTMObjects["HTM1"].layers["SensoryLayer"].activeColumns ))
+      # print("WINNERCELLS:"+str(serverData.HTMObjects["HTM1"].layers["SensoryLayer"].winnerCells))
+      # print("ACTIVECELLS:" + str(serverData.HTMObjects["HTM1"].layers["SensoryLayer"].activeCells))
+      # print("PREDICTCELLS:"+str(serverData.HTMObjects["HTM1"].layers["SensoryLayer"].predictiveCells))
+
+      pandaServer.serverData = serverData
+
+      pandaServer.spatialPoolers["HTM1"] = self.sp
+      pandaServer.temporalMemories["HTM1"] = self.tm
+      pandaServer.NewStateDataReady()
+
+
+import pandas as pd
+import os.path as path
+from nab.corpus import Corpus
+dataDir =  path.abspath(path.join(__file__ ,"../../../..","data"))
+
+corpus = Corpus(dataDir)
+
+dataSet = corpus.dataFiles["artificialWithAnomaly/art_daily_flatmiddle.csv"]
+
+detector = HtmcoreDetector(dataSet=dataSet,
+                probationaryPercent=0.15)
+
+detector.initialize()
+
+detector.run()
